0028 std_string_size の戻り値型を int から size_t に変更する
- webrtc/src/webrtc_c/std.h: 宣言の戻り値型を int → size_t
- webrtc/src/webrtc_c/std.cc: 実装の static_cast<int> を撤去し size_t で返す
- src/cxxstd.rs: CxxStringRef::len() の ffi 戻り値型変更に追従し len.max(0) as usize を撤去
- CHANGES.md: develop に [CHANGE] エントリを追記

+++ webrtc/src/webrtc_c/std.cc
@@ -16,9 +16,9 @@
 extern "C" {
 WEBRTC_DEFINE_UNIQUE(std_string, std::string);
 
-WEBRTC_EXPORT int std_string_size(struct std_string* self) {
+WEBRTC_EXPORT size_t std_string_size(struct std_string* self) {
   auto str = reinterpret_cast<std::string*>(self);
-  return static_cast<int>(str->size());
+  return str->size();
 }
 WEBRTC_EXPORT const char* std_string_c_str(struct std_string* self) {
   auto str = reinterpret_cast<std::string*>(self);
